refactor: add validation

+++ src/data/services/AllocationPeriod/DeleteAllocationPeriod.service.js
@@ -1,8 +1,14 @@
-import { AllocationPeriodRepository } from '../../database/repositories';
+import {
+  AllocationPeriodRepository,
+  AllocationRepository,
+  ProductHistoryRepository,
+} from '../../database/repositories';
 
 export class DeleteAllocationPeriodService {
   async execute({ id_allocation_period }) {
     const repository = new AllocationPeriodRepository();
+    const allocationRepository = new AllocationRepository();
+    const productHistoryRepository = new ProductHistoryRepository();
 
     const verifyAllocationPeriodExists = await repository.findAllocationPeriodById(
       {
@@ -14,6 +20,32 @@ export class DeleteAllocationPeriodService {
       return {
         error: `Não há nenhum Período de Alocação registrado com este ID -> ${id_allocation_period}.`,
       };
+
+    const verifyFkAllocation = await allocationRepository.verifyRelationAllocationPeriod(
+      {
+        id_allocation_period,
+      }
+    );
+
+    if (verifyFkAllocation.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Período de Alocação pois existem Alocações associadas.',
+      };
+    }
+
+    const verifyFkProductHistory = await productHistoryRepository.verifyRelationAllocationPeriod(
+      {
+        id_allocation_period,
+      }
+    );
+
+    if (verifyFkProductHistory.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Período de Alocação pois existem Históricos de produtos associados.',
+      };
+    }
 
     await repository.deleteAllocationPeriod({
       id_allocation_period,
